Fix issue with govendor
Govendor puts packages into a vendor subdirectory that creates folders
based on the url of the package. Since a bunch of packages are from github.com,
this ignore file entry filtered out the entire github.com subdir of the vendor
directory, so newly vendored files had to be force added.