Set backupcopy=yes for vim
Vim's default behaviour when updating backup copies is to rename the
original file to the name of the backup file (file.txt to file.txt~),
then create a new version of the original file (file.txt) and write the
file contents to it.

This breaks file watchers that watch via inodes. Setting backupcopy to
yes changes the backup behaviour to the following:
 - the contents of the original file (file.txt) are copied to the backup
   file (file.txt~)
 - the file contents are then written to the original file

This doesn't break file watchers.