Add ascend directory completion
Adds a 'dc' command (the reverse of cd). When the command is given an
argument, it works the same as cd. When used without an argument, it
just does 'cd ..'. When used with no arguments and tab is pressed, it
fills in the parent directory. Every subsequent time tab is pressed, the
auto-completion removes one level of the directory tree, so that you can
quickly jump up a directory again and again.